Fix: slot and votedAt are now strings

+++ backend-services/voting-app/src/main/java/org/cardano/foundation/voting/service/metadata/MetadataProcessor.java
@@ -57,7 +57,7 @@ public class MetadataProcessor {
     private JsonService jsonService;
 
     @Value("${l1.transaction.metadata.label:12345}")
-    private int metadataLabel;
+    private long metadataLabel;
 
     public void processMetadataEvents(List<TransactionMetadataLabelCbor> onChainMetadataEvents) {
         log.info("On chain events:{}", onChainMetadataEvents);
